pyproject.toml
[tool.poetry]
name = "django_planemodel"
version = "0.1.0"
description = ""
authors = ["SharProd <Sasha.sharai1998@mail.ru>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "4.0.5"
django-mptt = "0.13.4"
django-mptt-admin = "2.2.0"
Pillow = "9.1.1"
pytils = "0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
